+++ Software/src/LedDeviceLightpack.hpp
@@ -62,9 +62,9 @@ public slots:
     virtual void setColorSequence(QString /*value*/);
     virtual void requestFirmwareVersion();
     virtual void updateDeviceSettings();
-    virtual size_t maxLedsCount();
-    virtual size_t defaultLedsCount() { return maxLedsCount(); }
-    size_t lightpacksFound() { return m_devices.size(); }
+	virtual int maxLedsCount();
+	virtual int defaultLedsCount() { return maxLedsCount(); }
+	int lightpacksFound() { return m_devices.size(); }
 
 private: 
     bool readDataFromDevice();
